requirements.txt
langchain
langchain-core
langchain-community
langgraph
langsmith
python-dotenv
google-genai
langchain-google-genai
ipython
pillow
fastapi
uvicorn